requirements.txt
torch>=2.1
transformers>=4.40
datasets>=2.19
evaluate>=0.4.2
accelerate>=0.30
peft>=0.11.1
jiwer>=3.0.3
soundfile>=0.12.1
librosa>=0.10.2
pyyaml>=6.0.1
tqdm>=4.66

# Optional (only for word-level timestamps)
whisperx>=3.1.1
